requirements.txt
# 圖片處理工具依賴
Pillow>=10.0.0
numpy>=1.24.0

# PyQt6 - GUI 框架
PyQt6>=6.6.0
